Migrate builder tool to maven
