+++ .gitignore
@@ -1,2 +1,5 @@
 *.class
 **/.vscode/
+.gradle
+build
+bin
