+++ builder/src/main/java/rhit/Main.java
@@ -1,0 +1,7 @@
+package rhit;
+
+public class Main {
+  public static void main(String[] args) {
+    new BuilderInterface();
+  }
+}
